Include intermediate paths of explicitly included files
When a path is explicitly included, all its ancestors are included
without implying inclusion of their descendants. This way, if you
include something under an excluded directory, the ancestors are
tracked. As discussed in README.md and filter.go, this doesn't work
for pattern and base inclusions because we don't know the list of
matched paths in advance.