Bug fix: avoid segfault if Lstat fails
This was not previously exercised (and a comment even said so), but
the code was wrong and would segfault. Now it prints an error and
ignores the file. The situation of not being able to stat a file that
was previously seen during directory traversal can occur if there is a
disconnected fuse file system.